backend/requirements.txt
# FastAPI & Server
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.30
alembic==1.13.1

# ML & Data (Optional fallback)
# scikit-learn
# pandas
# numpy
joblib
anthropic

# NLP
nltk==3.8.1

# OCR & File Parsing
pytesseract==0.3.10
opencv-python-headless==4.9.0.80
Pillow==10.3.0
pdf2image==1.17.0
python-docx==1.1.2
pypdf==4.2.0

# PDF generation
reportlab==4.1.0

# Rate limiting
slowapi==0.1.9

# HTTP & CORS
httpx==0.27.0

# Environment
python-dotenv==1.0.1

# Validation
pydantic[email]==2.7.1
pydantic-settings==2.2.1

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
